Finalize writing
-Add ToString() to plane
-Make UVAxis use Vertex instead of storing individual points
-Update DispRows to use proper values instead of strings
-Make Vertex store points as doubles instead of integers
-Fix VClassReader not adding hidden classes & groups to SubClasses
-Fix VClassReader not parsing hammer++ vertices properly
-Fix VClassReader not assigning editor & hidden properties to Entities
-Expand upon VClassWriter's writing capabilities(so now it can properly write entities, solids, etc)
-Make Connections store blank properties as blank properties instead of null
-Add Hidden property to Entity
-Fix VerticesPlus not having a proper header
-Why would you trim parsed[1] so many times?
-Add blank constructor for VProperty

+++ Objects/vec3.cs
@@ -36,19 +36,19 @@
     
     public class Vertex
     {
-        public int X;
-        public int Y;
-        public int Z;
+        public double X;
+        public double Y;
+        public double Z;
         
         public Vertex(string str)
         {
             var property = str.Trim('[', ']').Split(' ');
-            X = int.Parse(property[0]);
-            Y = int.Parse(property[1]);
-            Z = int.Parse(property[2]);
+            X = double.Parse(property[0]);
+            Y = double.Parse(property[1]);
+            Z = double.Parse(property[2]);
         }
 
-        public Vertex(int x, int y, int z)
+        public Vertex(double x, double y, double z)
         {
             X = x;
             Y = y;
